+++ tests/repository/test_publish_container.py
@@ -36,7 +36,7 @@ def test_publish_order(requests_mocker, client):
     )
 
     # It should have succeeded, with the tasks as retrieved from Pulp
-    assert sorted(repo.publish().result()) == [
+    assert sorted(repo.publish()) == [
         Task(id="task1", succeeded=True, completed=True),
         Task(id="task2", succeeded=True, completed=True),
         Task(id="task3", succeeded=True, completed=True),
@@ -86,7 +86,7 @@ def test_publish_origin_only(requests_mocker, client):
     )
 
     # It should have succeeded, with the tasks as retrieved from Pulp
-    assert sorted(repo.publish(PublishOptions(origin_only=True)).result()) == [
+    assert sorted(repo.publish(PublishOptions(origin_only=True))) == [
         Task(id="task1", succeeded=True, completed=True),
         Task(id="task2", succeeded=True, completed=True),
     ]